requirements.txt
pytube
tqdm
pyJson
librosa
gnutools-python
nagisa